feat(carousel): Callbacks when carousel transitions begin and end.
This commit adds callbacks for the begin and end of the transition animations.
When a transition starts, the begin callbacks are invoked (entering or leaving versions).
When the transitioning ends, the end callbacks are invoked (entering or leaving versions).

This introduces the following attributes on the slide directive:
	onenteringtransitionbegin
	onenteringtransitionend
	onleavingtransitionbegin
	onleavingtransitionend

These will cause a callback to be before (within the scope).

Example:

<slide ... onenteringtransitionbegin="enteringEnded()" ...>
Where the scope has the following function:  $scope.enteringEnded = function enteringEnded() { ... };